move update_every and skip_noisy to optimizer
