+++ xnmt/optimizer.py
@@ -4,6 +4,7 @@ import numpy as np
 from xnmt import logger
 from xnmt.param_collection import ParamManager
 from xnmt.persistence import serializable_init, Serializable
+from xnmt import util
 
 """
 The purpose of this module is mostly to expose the DyNet trainers to YAML serialization,
@@ -13,12 +14,35 @@ but may also be extended to customize optimizers / training schedules
 class XnmtOptimizer(object):
   """
   A base classe for trainers. Trainers are mostly simple wrappers of DyNet trainers but can add extra functionality.
+
+  Args:
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
+  def __init__(self, update_every: int = 1, skip_noisy: bool = False) -> None:
+    self.update_every = update_every
+    self.num_updates_ignored = 0
+    self.skip_noisy = skip_noisy
+    if skip_noisy:
+      self.rolling_stats = util.RollingStatistic()
+
   def update(self):
     """
     Update the parameters.
     """
-    return self.optimizer.update()
+    self.num_updates_ignored += 1
+    if self.num_updates_ignored == self.update_every:
+      if not (self.skip_noisy and self._check_gradients_noisy()):
+        self.optimizer.update()
+      else:
+        logger.info("skipping noisy update")
+      self.num_updates_ignored = 0
+    else:
+      assert 0 < self.num_updates_ignored < self.update_every
 
   def status(self):
     """
@@ -63,6 +87,21 @@ class XnmtOptimizer(object):
   def learning_rate(self, value):
       self.optimizer.learning_rate = value
 
+  def _check_gradients_noisy(self) -> bool:
+    sq_norm = 0
+    for subcol in ParamManager.param_col.subcols.values():
+      for param in subcol.parameters_list():
+        cur_grads = param.grad_as_array()
+        sq_norm += np.sum(np.square(cur_grads))
+    log_norm = np.log(np.sqrt(sq_norm))
+    self.rolling_stats.update(log_norm)
+    if self.rolling_stats.average is None: # too few statistics
+      return False
+    else:
+      req_min = self.rolling_stats.average - 4*self.rolling_stats.stddev
+      req_max = self.rolling_stats.average + 4*self.rolling_stats.stddev
+      return not (req_min < log_norm < req_max)
+
 class SimpleSGDTrainer(XnmtOptimizer, Serializable):
   """
   Stochastic gradient descent trainer
@@ -71,11 +110,18 @@ class SimpleSGDTrainer(XnmtOptimizer, Serializable):
 
   Args:
     e0 (number): Initial learning rate
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!SimpleSGDTrainer'
 
   @serializable_init
-  def __init__(self, e0 = 0.1):
+  def __init__(self, e0 = 0.1, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.SimpleSGDTrainer(ParamManager.global_collection(), e0)
 class MomentumSGDTrainer(XnmtOptimizer, Serializable):
   """
@@ -86,11 +132,18 @@ class MomentumSGDTrainer(XnmtOptimizer, Serializable):
   Args:
     e0 (number): Initial learning rate
     mom (number): Momentum
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!MomentumSGDTrainer'
 
   @serializable_init
-  def __init__(self, e0 = 0.01, mom = 0.9):
+  def __init__(self, e0 = 0.01, mom = 0.9, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.MomentumSGDTrainer(ParamManager.global_collection(), e0, mom)
 
 class AdagradTrainer(XnmtOptimizer, Serializable):
@@ -102,11 +155,18 @@ class AdagradTrainer(XnmtOptimizer, Serializable):
   Args:
     e0 (number): Initial learning rate
     eps (number): Epsilon parameter to prevent numerical instability
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!AdagradTrainer'
 
   @serializable_init
-  def __init__(self, e0 = 0.1, eps = 1e-20):
+  def __init__(self, e0 = 0.1, eps = 1e-20, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.AdagradTrainer(ParamManager.global_collection(), e0, eps=eps)
 
 class AdadeltaTrainer(XnmtOptimizer, Serializable):
@@ -118,11 +178,18 @@ class AdadeltaTrainer(XnmtOptimizer, Serializable):
   Args:
     eps (number): Epsilon parameter to prevent numerical instability
     rho (number): Update parameter for the moving average of updates in the numerator
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!AdadeltaTrainer'
 
   @serializable_init
-  def __init__(self, eps = 1e-6, rho = 0.95):
+  def __init__(self, eps = 1e-6, rho = 0.95, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.AdadeltaTrainer(ParamManager.global_collection(), eps, rho)
 
 class AdamTrainer(XnmtOptimizer, Serializable):
@@ -136,11 +203,18 @@ class AdamTrainer(XnmtOptimizer, Serializable):
     beta_1 (number): Moving average parameter for the mean
     beta_2 (number): Moving average parameter for the variance
     eps (number): Epsilon parameter to prevent numerical instability
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!AdamTrainer'
 
   @serializable_init
-  def __init__(self, alpha = 0.001, beta_1 = 0.9, beta_2 = 0.999, eps = 1e-8):
+  def __init__(self, alpha = 0.001, beta_1 = 0.9, beta_2 = 0.999, eps = 1e-8, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.AdamTrainer(ParamManager.global_collection(), alpha, beta_1, beta_2, eps)
 
 class TransformerAdamTrainer(XnmtOptimizer, Serializable):
@@ -155,11 +229,18 @@ class TransformerAdamTrainer(XnmtOptimizer, Serializable):
     beta_1 (float):
     beta_2 (float):
     eps (float):
+    update_every: accumulate gradients for n minibatches before updating network parameters (mathematically equivalent
+                  to increasing batch size in case of ``sum`` loss_comb_method, but saves memory at the expense of
+                  slower computation).
+    skip_noisy: keep track of a moving average and a moving standard deviation of the log of the gradient norm
+                          values, and abort a step if the norm of the gradient exceeds four standard deviations of the
+                          moving average. Reference: https://arxiv.org/pdf/1804.09849.pdf
   """
   yaml_tag = '!TransformerAdamTrainer'
 
   @serializable_init
-  def __init__(self, alpha=1.0, dim=512, warmup_steps=4000, beta_1=0.9, beta_2=0.98, eps=1e-9):
+  def __init__(self, alpha=1.0, dim=512, warmup_steps=4000, beta_1=0.9, beta_2=0.98, eps=1e-9, update_every: int = 1, skip_noisy: bool = False):
+    super().__init__(update_every=update_every, skip_noisy=skip_noisy)
     self.optimizer = dy.AdamTrainer(ParamManager.global_collection(),
                                     alpha=alpha,
                                     beta_1=beta_1,
